refactor, added commentWrapperm and ifBinding
