+++ src/ifBinding.js
@@ -1,0 +1,13 @@
+// import * as config from './config';
+// import * as util from './util';
+// import {insertRenderedElements} from './commentWrapper';
+
+const renderIfBinding = ({bindingData, viewModel, bindingAttrs}) => {
+    // TODO: render element
+};
+
+const removeIfBinding = ({bindingData, viewModel, bindingAttrs}) => {
+    // TODO: render element
+};
+
+export {renderIfBinding, removeIfBinding};
